Make the subroutines in the two ctsm modules have different interfaces
This still avoids the linker error, but we still use the wrong one

+++ shr_kind_dependency/atm_lib/ctsm.f90
@@ -11,13 +11,13 @@ module ctsm
 contains
 
   subroutine ctsm_in(x)
-    real*8, intent(in) :: x
+    integer, intent(in) :: x
 
     print *, "atmosphere's version: ", x
   end subroutine ctsm_in
 
   subroutine ctsm_out(x)
-    real*8, intent(out) :: x
+    integer, intent(out) :: x
 
     x = 170
   end subroutine ctsm_out
